Consolidate documentation and tests: Reduce PR files from 16 to 9
- Merged 5 separate documentation files into COMPLETE_DOCUMENTATION.md
- Merged 3 separate test files into test_all_languages_comprehensive.py
- Removed redundant documentation files
- All 32 test cases now in single file for easy review
- Installation guide, quick start, and test reports all in one place
