--- JAVASCRIPT_TEST_RESULTS.md
@@ -1,156 +1,0 @@
-# JavaScript Complex Test Results
-
-## Test Summary
-All 12 complex JavaScript tests passed successfully! ✅
-
-### Test Results Breakdown:
-
-#### ✅ Test 1: Unhandled Promise Rejection
-- **Status**: PASSED
-- **Result**: Correctly caught error and exited with code 1
-- **Error captured**: "Network timeout!"
-
-#### ✅ Test 2: Promise Chain Error Propagation
-- **Status**: PASSED
-- **Result**: Error propagated correctly through promise chain
-- **Output**: Showed steps 1-2, then caught error at step 2
-
-#### ✅ Test 3: Infinite Loop (Timeout Protection)
-- **Status**: PASSED
-- **Result**: Timeout triggered after 15 seconds as expected
-- **Protection**: Prevents hanging on infinite loops
-
-#### ✅ Test 4: Stack Overflow
-- **Status**: PASSED
-- **Result**: Stack overflow caught and error message displayed
-- **Output**: Showed recursion depth up to ~460+ before overflow
-- **Error captured**: "Maximum call stack size exceeded"
-
-#### ✅ Test 5: Type Coercion Bugs
-- **Status**: PASSED
-- **Result**: Demonstrated JavaScript's quirky type coercion
-- **Interesting outputs**:
-  - `[] + [] = ""` (empty string)
-  - `[] + {} = "[object Object]"`
-  - `'5' - 3 = 2` (subtraction coerces to number)
-  - `'5' + 3 = "53"` (addition concatenates)
-  - `NaN === NaN = false`
-
-#### ✅ Test 6: Closure Memory Pattern
-- **Status**: PASSED
-- **Result**: Closures correctly captured individual data
-- **Output**: Each closure maintained its own scope
-
-#### ✅ Test 7: Event Loop & Microtasks
-- **Status**: PASSED
-- **Result**: Correct execution order demonstrated
-- **Order**: Synchronous → Microtasks → Macrotasks
-  1. Synchronous start
-  2. Synchronous end
-  3. Promise (microtask)
-  4. queueMicrotask (microtask)
-  5. setTimeout (macrotask)
-
-#### ✅ Test 8: Regex with Potential Backtracking
-- **Status**: PASSED
-- **Result**: Regex executed in 87ms (acceptable performance)
-- **Pattern**: `/^(a+)+$/` on "aaaaaaaaaaaaaaaaaaaaX"
-- **Result**: false (correctly detected non-match)
-
-#### ✅ Test 9: Async/Await Error Handling
-- **Status**: PASSED
-- **Result**: Error propagated correctly through async layers
-- **Error chain**: complexAsyncFlow → main
-- **Caught and re-thrown**: "Re-throwing: Operation failed!"
-
-#### ✅ Test 10: Object Property Edge Cases
-- **Status**: PASSED
-- **Result**: Handled edge cases correctly
-- **Features tested**:
-  - Undefined property access
-  - Safe navigation operator (`?.`)
-  - Empty string keys
-  - Numeric string keys
-  - Object.create(null) usage
-
-#### ✅ Test 11: Large Output (Truncation)
-- **Status**: PASSED
-- **Result**: Output correctly truncated at 5000 chars
-- **Generated**: 1000 lines of 50 characters each
-- **Protection**: Prevents memory overflow from massive outputs
-
-#### ✅ Test 12: Syntax Error
-- **Status**: PASSED
-- **Result**: Syntax error caught with detailed stack trace
-- **Error**: "missing ) after argument list"
-- **Exit code**: 1 (correctly indicates failure)
-
----
-
-## Key Features Verified:
-
-### 1. Error Handling ✅
-- Promise rejections
-- Async/await errors
-- Syntax errors
-- Runtime errors
-- Stack overflow
-
-### 2. Timeout Protection ✅
-- 15-second timeout working
-- Prevents infinite loops
-- Protects against hangs
-
-### 3. Output Management ✅
-- Captures stdout
-- Captures stderr
-- Truncates at 5000 characters
-- Handles large outputs gracefully
-
-### 4. JavaScript-Specific Features ✅
-- Async/await
-- Promises
-- Event loop behavior
-- Microtasks vs Macrotasks
-- Closures
-- Type coercion
-- Regex execution
-
-### 5. Exit Code Tracking ✅
-- Successful execution: 0
-- Errors: 1
-- Properly tracked and reported
-
----
-
-## Performance Metrics:
-
-| Test | Execution Time | Status |
-|------|---------------|--------|
-| Promise Rejection | < 1s | ✅ Fast |
-| Promise Chain | < 1s | ✅ Fast |
-| Infinite Loop | 15s (timeout) | ✅ Protected |
-| Stack Overflow | < 1s | ✅ Fast |
-| Type Coercion | < 1s | ✅ Fast |
-| Closures | < 1s | ✅ Fast |
-| Event Loop | < 1s | ✅ Fast |
-| Regex | 87ms | ✅ Fast |
-| Async/Await | < 1s | ✅ Fast |
-| Object Props | < 1s | ✅ Fast |
-| Large Output | < 1s | ✅ Fast |
-| Syntax Error | < 1s | ✅ Fast |
-
----
-
-## Conclusion:
-
-The JavaScript executor implementation is **production-ready** and handles:
-- ✅ All error types (syntax, runtime, async)
-- ✅ Timeout protection (15s limit)
-- ✅ Output truncation (5000 char limit)
-- ✅ Complex async patterns
-- ✅ JavaScript quirks and edge cases
-- ✅ Stack overflow protection
-- ✅ Large output handling
-
-**Overall Assessment**: Excellent! The implementation is robust, secure, and handles edge cases gracefully.
