--- TEST_SUITE_SUMMARY.md
@@ -1,151 +1,0 @@
-# Test Suite Summary
-
-This document provides a quick reference for all test files used to validate the multi-language implementation.
-
-## Test Files Location
-
-All test files are organized in the `tests/` directory:
-
-```
-tests/
-├── README.md                              # Comprehensive test documentation
-├── test_complex_new_languages.py          # Stress tests for 5 languages (Rust, Ruby, PHP, C, C#)
-├── test_complex_new_languages_success.py  # Success validation for 5 languages
-└── test_complex_js.py                     # JavaScript comprehensive tests (12 scenarios)
-```
-
-## Quick Test Commands
-
-### Run All Tests (Sequential)
-```bash
-# From project root directory
-python tests\test_complex_new_languages.py
-python tests\test_complex_new_languages_success.py
-python tests\test_complex_js.py
-```
-
-### Run Individual Test Suites
-```bash
-# Stress tests for new languages
-python tests\test_complex_new_languages.py
-
-# Success validation tests
-python tests\test_complex_new_languages_success.py
-
-# JavaScript complex tests
-python tests\test_complex_js.py
-```
-
-## Test Coverage Summary
-
-| Language | Stress Tests | Success Tests | Total |
-|----------|--------------|---------------|-------|
-| Rust     | 3            | 1             | 4     |
-| Ruby     | 3            | 1             | 4     |
-| PHP      | 3            | 1             | 4     |
-| C        | 3            | 1             | 4     |
-| C#       | 3            | 1             | 4     |
-| JavaScript | 12         | N/A           | 12    |
-| **Total** | **27**      | **5**         | **32** |
-
-## Test Statistics
-
-- **Total Test Files**: 3
-- **Total Test Scenarios**: 32
-- **Languages Tested**: 6 (JavaScript, Rust, Ruby, PHP, C, C#)
-- **Estimated Total Runtime**: 5-7 minutes
-- **All Tests Status**: ✅ PASSING
-
-## What Each Test File Validates
-
-### 1. test_complex_new_languages.py
-- **Purpose**: Stress testing and edge case validation
-- **Validates**: Error handling, timeout protection, output truncation, stack overflow detection
-- **Test Count**: 15 tests (3 per language)
-
-### 2. test_complex_new_languages_success.py
-- **Purpose**: Correct program execution validation
-- **Validates**: Successful compilation and execution of working programs
-- **Test Count**: 5 tests (1 per language)
-
-### 3. test_complex_js.py
-- **Purpose**: JavaScript-specific feature testing
-- **Validates**: Async/await, promises, event loop, type coercion, closures, error handling
-- **Test Count**: 12 comprehensive scenarios
-
-## Prerequisites Check
-
-Before running tests, verify all dependencies are installed:
-
-```bash
-node --version    # Should show v22.19.0+
-rustc --version   # Should show 1.90.0+
-ruby --version    # Should show 3.4.7+
-php --version     # Should show 8.4.13+
-dotnet --version  # Should show 9.0.305+
-gcc --version     # Should show MinGW version
-```
-
-If any are missing, refer to `INSTALLATION_GUIDE.md` for installation instructions.
-
-## Test Organization Rationale
-
-The tests are organized to separate concerns:
-
-1. **Stress Tests** (`test_complex_new_languages.py`):
-   - Tests error conditions
-   - Validates safety mechanisms (timeout, truncation)
-   - Ensures robustness under extreme conditions
-
-2. **Success Tests** (`test_complex_new_languages_success.py`):
-   - Tests correct program execution
-   - Validates language-specific features
-   - Ensures proper output generation
-
-3. **JavaScript Tests** (`test_complex_js.py`):
-   - JavaScript-specific testing
-   - Validates async patterns
-   - Tests Node.js runtime behavior
-
-## Files NOT Included in tests/
-
-The following test files were used during development but are NOT part of the final test suite:
-
-- `test_new_languages.py` - Initial development tests for Rust/Ruby/PHP
-- `test_c_csharp.py` - Initial development tests for C/C#
-- Various temporary test files created during development
-
-These files remain in the root directory for reference but are not part of the official test suite.
-
-## Running Tests from VS Code
-
-1. Open integrated terminal (`` Ctrl+` ``)
-2. Ensure you're in the project root
-3. Run test commands as shown above
-
-## Continuous Integration
-
-These tests can be integrated into CI/CD pipelines:
-
-```yaml
-# Example GitHub Actions workflow
-- name: Run Test Suite
-  run: |
-    python tests/test_complex_new_languages.py
-    python tests/test_complex_new_languages_success.py
-    python tests/test_complex_js.py
-```
-
-## Test Maintenance
-
-When adding new languages:
-1. Add stress tests to `test_complex_new_languages.py`
-2. Add success validation to `test_complex_new_languages_success.py`
-3. Update test counts in this document
-4. Update `tests/README.md` with new test descriptions
-
----
-
-**Last Updated**: October 13, 2025  
-**Test Suite Version**: 1.0  
-**Languages Covered**: 6 (JavaScript, Rust, Ruby, PHP, C, C#)
